refactoring: separate element mixins from shady mutations
